+++ base_proc/modules/rdinv.py
@@ -1,8 +1,7 @@
 #!../.wenv_base_proc/bin/python3
 
-import pylightxl as xl
-from datetime import datetime
-from colorama import Fore, Back, Style
+
+
 
 def rdinv(file_to_process: str):
     """ RDINV - modul de procesare a fisierului format XLSX ce contine factura si colectare a datelor aferente
@@ -17,13 +16,27 @@ def rdinv(file_to_process: str):
         INTRARI: fisier format XLSX ce contine factura emisa (cod: `f-XLSX`)
         IESIRI: fisier format JSON imagine a datelor facturii (cod: `f-JSON`)
     """
+    # general modules (should be already imported by BaseProc class)
+    from datetime import datetime
+    from colorama import Fore, Back, Style
 
     import pylightxl as xl
 
+    print(f"\n*** Module {Fore.RED} RDINV (code-name: `rdinv`) {Style.RESET_ALL} started at {Fore.GREEN} {datetime.now()} {Style.RESET_ALL} to process file {Fore.GREEN} {file_to_process} {Style.RESET_ALL}")
 
-    print(f"*** Module `rdinv` started at {Fore.GREEN} {datetime.now()} {Style.RESET_ALL} to process file {Fore.GREEN} {file_to_process} {Style.RESET_ALL}")
+    # read Excel file, the worksheed with Invoice data
+    db = xl.readxl(fn=file_to_process)
+    print(f"DEBUG-note: Excel database read (`db` variable) as: {db}") #NOTE for debug purposes
+    invoice_worksheet_name = "FACTURA FINALA" #TODO workshhet ("FACTURA FINALA" now) should be subject of configuration
+    ws = db.ws(invoice_worksheet_name)
+    print(f"DEBUG-note: Excel worksheet read (`ws` variable) as: {ws}") #NOTE for debug purposes
 
-    ... #TODO my code here
+    #FIXME for test read cell B13 which should contain text "Cota TVA: 19%"
+    tmp = ws.address(address="B13")
+    print(f"---> TEST-note: cell read contains: {tmp}")
+    #FIXME -------------------------------------------------------------------------------------- END OF TEST here {#NOTEOK-PASSED}
+
+    ... #TODO ...hereuare...
     return None